requirements.txt
streamlit
python-dotenv
Pillow
google-generativeai
